various adjustments.
1. Needed to take newing up the ReflowOperationState out of the
SetupHandler. This was causing weirdness, probably because it needed a
copy constructor but i didn't give it one because newing stuff up should
only happen in one place anyway.

2. Implement a method for printing out the reflow state.

3. Improve how we deal with figuring out target DTdt. Now it's just
doing division at runtime instead of storing those things in an array.
If this is a performance issue later I'll deal with it then.

4. renaming targetDTdt methods as such.