a bunch of c++ updates
This includes a bunch of simplifications to reflowOperationState and
loopHandler.

We are also simplifying the serial printing a bit, since we are now
wanting to read to a CSV file using Processing.

TODO: There is still a bit of instability in the system w/r/t strange
serial output. I'm not sure but I wonder if there are buffer overflows
in play.
.
There is some question of how to use the PID output and what to compare
it with.

it's not great but the commit has gotten too big to backtrack.